Backend/requirements.txt
phidata
pypdf
psycopg2-binary
sqlalchemy
python-dotenv
fastapi
uvicorn
groq
google-search-results
fastapi
uvicorn
python-dotenv

